ref: moved common time utils to utils dir
