+++ rocketry/core/task.py
@@ -28,7 +28,7 @@ from rocketry.core.condition import BaseCondition, AlwaysFalse, All
 from rocketry.core.time import TimePeriod
 from rocketry.core.parameters import Parameters
 from rocketry.core.log import TaskAdapter
-from rocketry.core.time.utils import to_timedelta
+from rocketry.pybox.time import to_timedelta
 from rocketry.core.utils import is_pickleable, filter_keyword_args, is_main_subprocess
 from rocketry.exc import SchedulerRestart, SchedulerExit, TaskInactionException, TaskTerminationException, TaskLoggingError, TaskSetupError
 from rocketry.core.hook import _Hooker
