+++ rocketry/pybox/time/__init__.py
@@ -1,2 +1,6 @@
-from .convert import to_microseconds, to_timedelta, to_datetime
+from .convert import (
+    to_microseconds, to_timedelta, to_datetime,
+    timedelta_to_str, datetime_to_dict, to_timedelta,
+    to_timestamp
+)
 from .interval import Interval
